step 1: extract sorted_internals.rs
Pure relocation of the binary_search_null_boundary and
binary_search_sorted helpers from range_query_fastfield.rs into a new
sibling module src/query/range_query/sorted_internals.rs, marked
pub(crate). Bodies are unchanged - the DESC monotonicity invariant
that range_query's sorted path depends on is preserved exactly.

Step 1. The term-set gallop strategy
(Step 3) reuses these helpers without duplicating numeric logic.
